Added more flexibility, improved support for scale handling
Separated the inset calculation as it alone was needed for my summer
research project, but other changes I made in order to better support
retina displays seemed worthy of contributing back.

Now an optional boolean flag can be passed to allow the image to be
cropped based on whether pixels are semi-transparent. See the demo
Xcode project for an example.